+++ notification/test/integration/init/init-resources.js
