--- notification/test/integration/init/initInterfaceInteractionCustomType.js
@@ -1,17 +1,0 @@
-const Promise = require('bluebird')
-
-const interfaceInteractionTypes = require('../../resources/payment-interface-interaction-types')
-
-async function initInterfaceInteractionCustomType (ctpClient) {
-  await Promise.map(interfaceInteractionTypes, async (type) => {
-    try {
-      await ctpClient.create(ctpClient.builder.types, type)
-    } catch (e) {
-      console.error('Error when creating interface interaction custom type, skipping...', JSON.stringify(e))
-    }
-  }, { concurrency: 3 })
-}
-
-module.exports = {
-  initInterfaceInteractionCustomType
-}
